requirements.txt
python-dotenv
openai
requests
python-telegram-bot
